# frontend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/games/{game_id}/board": {
      "get": {
        "tags": [
          "getBoard"
        ],
        "summary": "Board",
        "operationId": "getBoard",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "integer"
            },
            "name": "game_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidGameIdErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/move": {
      "post": {
        "tags": [
          "makeMove"
        ],
        "summary": "Make Move",
        "operationId": "makeMove",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "integer"
            },
            "name": "game_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidBoardIndexErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotUnavailableErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/newgame": {
      "post": {
        "tags": [
          "makeNewGame"
        ],
        "summary": "New Game",
        "operationId": "makeNewGame",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewGameResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidConnectionErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/settings": {
      "post": {
        "tags": [
          "makeSettings"
        ],
        "summary": "Make Settings",
        "operationId": "makeSettings",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "integer"
            },
            "name": "game_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidGameIdErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BoardResponse": {
        "title": "BoardResponse",
        "required": [
          "slots",
          "next_turn",
          "next_turn_token"
        ],
        "type": "object",
        "properties": {
          "slots": {
            "title": "Slots",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "next_turn": {
            "$ref": "#/components/schemas/Player"
          },
          "next_turn_token": {
            "title": "Next Turn Token",
            "type": "string"
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "InvalidBoardIndexErrorResponse": {
        "title": "InvalidBoardIndexErrorResponse",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "InvalidConnectionErrorResponse": {
        "title": "InvalidConnectionErrorResponse",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "InvalidGameIdErrorResponse": {
        "title": "InvalidGameIdErrorResponse",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "MoveRequest": {
        "title": "MoveRequest",
        "required": [
          "slot_index",
          "player"
        ],
        "type": "object",
        "properties": {
          "slot_index": {
            "title": "Slot Index",
            "type": "integer"
          },
          "player": {
            "$ref": "#/components/schemas/Player"
          }
        }
      },
      "NewGameResponse": {
        "title": "NewGameResponse",
        "required": [
          "game_id",
          "pin"
        ],
        "type": "object",
        "properties": {
          "game_id": {
            "title": "Game Id",
            "type": "integer"
          },
          "pin": {
            "title": "Pin",
            "type": "string"
          }
        }
      },
      "Player": {
        "title": "Player",
        "enum": [
          0,
          1,
          2
        ],
        "description": "An enumeration."
      },
      "SettingsRequest": {
        "title": "SettingsRequest",
        "required": [
          "player_one_token",
          "player_two_token"
        ],
        "type": "object",
        "properties": {
          "player_one_token": {
            "title": "Player One Token",
            "type": "string"
          },
          "player_two_token": {
            "title": "Player Two Token",
            "type": "string"
          }
        }
      },
      "SettingsResponse": {
        "title": "SettingsResponse",
        "required": [
          "response"
        ],
        "type": "object",
        "properties": {
          "response": {
            "title": "Response",
            "type": "integer"
          }
        }
      },
      "SpotUnavailableErrorResponse": {
        "title": "SpotUnavailableErrorResponse",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
      "ValidationErrorResponse": {
        "title": "ValidationErrorResponse",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Loc",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        }
      }
    }
  }
}